Completed video 37, end of Module 5. Completed deleteItem permissions on the server side

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -43,9 +43,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`)
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`)
     // check if they own that item, or have permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ["ADMIN", "ITEMDELETE"].includes(permission));
+
+    if (!ownsItem || !hasPermissions) {
+      throw new Error("You don't have permissions to do that!")
+    }
+
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
